+++ zarrs/examples/data_type_optional.rs
@@ -25,10 +25,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // Build the codec chains for the optional codec
     let array = ArrayBuilder::new(
-        vec![4, 4],                  // 4x4 array
-        vec![2, 2],                  // 2x2 chunks
-        DataType::UInt8.optional(),  // Optional uint8
-        FillValue::from(None::<u8>), // Null fill value: [0]
+        vec![4, 4],                      // 4x4 array
+        vec![2, 2],                      // 2x2 chunks
+        DataType::UInt8.into_optional(), // Optional uint8
+        FillValue::from(None::<u8>),     // Null fill value: [0]
     )
     .dimension_names(["y", "x"].into())
     .attributes(